silhouette-backend/requirements.txt
# API
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
pydantic-settings>=2.2.0

# LLM orchestration
langchain==0.1.12
langchain-community==0.0.28
langchain-groq==0.1.3
langgraph==0.0.26
langsmith==0.1.17

# Vector DB
chromadb==0.4.24

# Embeddings & Vision (HuggingFace)
transformers==4.38.2
torch==2.2.1
torchvision==0.17.1
Pillow==10.2.0
open-clip-torch==2.24.0

# Audio
faster-whisper>=1.1.0

# Retrieval
rank-bm25==0.2.2
scikit-learn==1.4.1.post1
numpy==1.26.4

# Ollama
ollama>=0.2.0

# Groq
groq>=0.4.2

# Utils
python-dotenv==1.0.1
aiofiles==23.2.1
httpx>=0.27.0